# unittests: Make it easier to determine on which _spec line it crashed